dev-lang/gnat-gpl: I just want a ada cross-compiler
Package-Manager: Portage-2.3.103, Repoman-2.3.23

+++ dev-lang/gnat-gpl/gnat-gpl-2019-r3.ebuild
@@ -93,7 +93,9 @@ src_unpack() {
 
 	toolchain_src_unpack
 	if use bootstrap; then
-		rm ${BTSTRP}/libexec/gcc/*/4.7.4/ld || die
+		if ! use arm; then
+			rm ${BTSTRP}/libexec/gcc/*/4.7.4/ld || die
+		fi
 	fi
 }
